requirements.txt
Mezzanine==4.2.3
django-embed-video==1.1.2
django-modeltranslation==0.12.1
psycopg2==2.7.3.2